requirements.txt
cassiopeia
discord.py
sqlalchemy
